chore(connect): move Fees classes
