+++ packages/connect/src/api/bitcoin/index.ts
@@ -1,6 +1,5 @@
 export * from './createPendingTx';
 export * from './enhanceSignTx';
-export * from './BitcoinFees';
 export * from './inputs';
 export * from './outputs';
 export * from './refTx';
