+++ packages/connect/src/backend/fees/__tests__/BitcoinFeeLevels.test.ts
@@ -2,9 +2,9 @@ import BlockchainLink from '@trezor/blockchain-link';
 import coinsJSONEth from '@trezor/connect-common/files/coins-eth.json';
 import coinsJSON from '@trezor/connect-common/files/coins.json';
 
-import { initBlockchain } from '../../../backend/BlockchainLink';
 import { getBitcoinNetwork, parseCoinsJson } from '../../../data/coinInfo';
-import { BitcoinFeeLevels } from '../BitcoinFees';
+import { initBlockchain } from '../../BlockchainLink';
+import { BitcoinFeeLevels } from '../BitcoinFeeLevels';
 
 describe('api/bitcoin/Fees', () => {
     // load coin definitions
